Add missing tool descriptions
Full credit due to @jhachirag7 on this, I've set them as the commit
author. They also opened #602 but it looks like it has some indentation
issue (fixed here) and a few spurious commits and removed files; I only
mean to make it as easy as possible for maintainers to get this merged.

I (mike@mbaynton.com) have tested that with this fix it is possible to
run against AWS Bedrock + Sonnet 3.5 v2.

Fixes #580